+++ src/main.cpp
@@ -16,6 +16,18 @@
 #include <fcntl.h>
 #endif
 
+#ifdef LSP_BUILD_WITH_SENTRY
+// sentry.h pulls in <windows.h>
+#ifdef _WIN32
+#define WIN32_LEAN_AND_MEAN
+#define NOMINMAX
+#endif
+#define SENTRY_BUILD_STATIC 1
+#include <sentry.h>
+#endif
+
+#define LSP_VERSION "1.46.0"
+
 static void displayFlags()
 {
     printf("Available flags:\n");
@@ -33,6 +45,31 @@ static void displayFlags()
 
 int startLanguageServer(const argparse::ArgumentParser& program)
 {
+    bool isCrashReportingEnabled = program.is_used("--enable-crash-reporting");
+    if (isCrashReportingEnabled)
+    {
+#ifdef LSP_BUILD_WITH_SENTRY
+        std::optional<std::filesystem::path> crashReportDirectory = program.present<std::filesystem::path>("--crash-report-directory");
+
+        sentry_options_t* options = sentry_options_new();
+        sentry_options_set_dsn(options, "https://bc658c75485d1aecbaf1c0c1f7980922@o4509305213026304.ingest.de.sentry.io/4509305221283920");
+
+        if (crashReportDirectory.has_value())
+        {
+#ifdef _WIN32
+            sentry_options_set_database_pathw(options, crashReportDirectory->c_str());
+#else
+            sentry_options_set_database_path(options, crashReportDirectory->c_str());
+#endif
+        }
+
+        sentry_options_set_release(options, "luau-lsp@" LSP_VERSION);
+        sentry_init(options);
+#else
+        std::cerr << "Ignoring '--enable-crash-reporting' as this server was not built with crash reporting features\n";
+#endif
+    }
+
     // Debug loop: set a breakpoint inside while loop to attach debugger before init
     if (program.is_used("--delay-startup"))
     {
@@ -117,6 +154,11 @@ int startLanguageServer(const argparse::ArgumentParser& program)
     // Begin input loop
     server.processInputLoop();
 
+#ifdef LSP_BUILD_WITH_SENTRY
+    if (isCrashReportingEnabled)
+        sentry_close();
+#endif
+
     // If we received a shutdown request before exiting, exit normally. Otherwise, it is an abnormal exit
     return server.requestedShutdown() ? 0 : 1;
 }
@@ -163,7 +205,7 @@ int main(int argc, char** argv)
         return 1;
     };
 
-    argparse::ArgumentParser program("luau-lsp", "1.46.0");
+    argparse::ArgumentParser program("luau-lsp", LSP_VERSION);
     program.set_assign_chars(":=");
 
     // Global arguments
@@ -253,6 +295,14 @@ int main(int argc, char** argv)
     lsp_command.add_argument("--stdio").help("set up client communication channel via stdio").implicit_value(true);
     lsp_command.add_argument("--pipe")
         .help("path to pipe / socket file name for pipe communication channel")
+        .action(file_path_parser)
+        .metavar("PATH");
+    lsp_command.add_argument("--enable-crash-reporting")
+        .help("whether to enable crash reporting to Sentry")
+        .default_value(false)
+        .implicit_value(true);
+    lsp_command.add_argument("--crash-report-directory")
+        .help("location to store database for crash reports")
         .action(file_path_parser)
         .metavar("PATH");