+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Implemented opt-in crash reporting for the language server. This is enabled via `luau-lsp.server.crashReporting.enabled`. Note: crash reporting
+  sends network requests to [Sentry](https://sentry.io/). Crash Reporting is only available on Windows and macOS, and is not enabled in Standalone mode (`luau-lsp analyze`).
+
 ### Changed
 
 - Sync to upstream Luau 0.674